Added newline to end of MetaData.cpp. Changed types in setInitialMJITCountUnsynchronized to prevent compiler warnings about increasing size. Fixed inconsitent placement of pointer specifier in variable declarations in CompilationThread.cpp

+++ runtime/compiler/control/CompilationRuntime.hpp
@@ -660,10 +660,10 @@ public:
 #if defined(J9VM_OPT_MICROJIT)
    static void setInitialMJITCountUnsynchronized(J9Method *method, int32_t mjitThreshold, int32_t trThreshold)
       {
-      int32_t value = 0;
+      intptr_t value = 0;
       if (mjitThreshold < trThreshold)
          {
-         value = ((trThreshold - mjitThreshold) << 1) | 1;
+         value = (intptr_t)(((trThreshold - mjitThreshold) << 1) | 1);
          }
       method->extra2 = reinterpret_cast<void *>(value);
       }
